Finish filters updating

+++ src/Plugins/SimplnxCore/test/SetImageGeomOriginScalingFilterTest.cpp
@@ -12,8 +12,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter(Instantiate)", "[Simplnx
   DataPath k_ImageGeomPath({Constants::k_SmallIN100, Constants::k_EbsdScanData, Constants::k_ImageGeometry});
   bool k_ChangeOrigin = false;
   bool k_ChangeResolution = false;
-  std::vector<float64> k_Origin{0, 0, 0};
-  std::vector<float64> k_Spacing{1, 1, 1};
+  std::vector<float32> k_Origin{0, 0, 0};
+  std::vector<float32> k_Spacing{1, 1, 1};
 
   SetImageGeomOriginScalingFilter filter;
   DataStructure dataStructure = UnitTest::CreateDataStructure();
@@ -22,8 +22,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter(Instantiate)", "[Simplnx
   args.insert(SetImageGeomOriginScalingFilter::k_SelectedImageGeometryPath_Key, std::make_any<DataPath>(k_ImageGeomPath));
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeOrigin_Key, std::make_any<bool>(k_ChangeOrigin));
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeSpacing_Key, std::make_any<bool>(k_ChangeResolution));
-  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float64>>(k_Origin));
-  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float64>>(k_Spacing));
+  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float32>>(k_Origin));
+  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float32>>(k_Spacing));
 
   auto result = filter.preflight(dataStructure, args);
   SIMPLNX_RESULT_REQUIRE_VALID(result.outputActions);
@@ -34,8 +34,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter: Valid Execution", "[Sim
   DataPath k_ImageGeomPath({Constants::k_SmallIN100, Constants::k_EbsdScanData, Constants::k_ImageGeometry});
   bool k_ChangeOrigin = true;
   bool k_ChangeResolution = true;
-  std::vector<float64> k_Origin{7, 6, 5};
-  std::vector<float64> k_Spacing{2, 2, 2};
+  std::vector<float32> k_Origin{7, 6, 5};
+  std::vector<float32> k_Spacing{2, 2, 2};
 
   SetImageGeomOriginScalingFilter filter;
   DataStructure dataStructure = UnitTest::CreateDataStructure();
@@ -44,8 +44,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter: Valid Execution", "[Sim
   args.insert(SetImageGeomOriginScalingFilter::k_SelectedImageGeometryPath_Key, std::make_any<DataPath>(k_ImageGeomPath));
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeOrigin_Key, std::make_any<bool>(k_ChangeOrigin));
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeSpacing_Key, std::make_any<bool>(k_ChangeResolution));
-  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float64>>(k_Origin));
-  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float64>>(k_Spacing));
+  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float32>>(k_Origin));
+  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float32>>(k_Spacing));
 
   auto preflightResult = filter.preflight(dataStructure, args);
   SIMPLNX_RESULT_REQUIRE_VALID(preflightResult.outputActions);
@@ -63,8 +63,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter: 0,0,0 Central Origin", 
   DataPath k_ImageGeomPath({Constants::k_SmallIN100, Constants::k_EbsdScanData, Constants::k_ImageGeometry});
   bool k_ChangeOrigin = true;
   bool k_ChangeResolution = true;
-  std::vector<float64> k_Origin{0.0, 0.0, 0.0};
-  std::vector<float64> k_Spacing{2, 2, 2};
+  std::vector<float32> k_Origin{0.0, 0.0, 0.0};
+  std::vector<float32> k_Spacing{2, 2, 2};
 
   SetImageGeomOriginScalingFilter filter;
   DataStructure dataStructure = UnitTest::CreateDataStructure();
@@ -74,8 +74,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter: 0,0,0 Central Origin", 
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeOrigin_Key, std::make_any<bool>(k_ChangeOrigin));
   args.insert(SetImageGeomOriginScalingFilter::k_CenterOrigin_Key, std::make_any<bool>(true));
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeSpacing_Key, std::make_any<bool>(k_ChangeResolution));
-  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float64>>(k_Origin));
-  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float64>>(k_Spacing));
+  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float32>>(k_Origin));
+  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float32>>(k_Spacing));
 
   auto preflightResult = filter.preflight(dataStructure, args);
   SIMPLNX_RESULT_REQUIRE_VALID(preflightResult.outputActions);
@@ -93,8 +93,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter: Custom Central Origin",
   DataPath k_ImageGeomPath({Constants::k_SmallIN100, Constants::k_EbsdScanData, Constants::k_ImageGeometry});
   bool k_ChangeOrigin = true;
   bool k_ChangeResolution = true;
-  std::vector<float64> k_Origin{7.0, 6.0, 5.0};
-  std::vector<float64> k_Spacing{2, 2, 2};
+  std::vector<float32> k_Origin{7.0, 6.0, 5.0};
+  std::vector<float32> k_Spacing{2, 2, 2};
 
   SetImageGeomOriginScalingFilter filter;
   DataStructure dataStructure = UnitTest::CreateDataStructure();
@@ -104,8 +104,8 @@ TEST_CASE("SimplnxCore::SetImageGeomOriginScalingFilter: Custom Central Origin",
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeOrigin_Key, std::make_any<bool>(k_ChangeOrigin));
   args.insert(SetImageGeomOriginScalingFilter::k_CenterOrigin_Key, std::make_any<bool>(true));
   args.insert(SetImageGeomOriginScalingFilter::k_ChangeSpacing_Key, std::make_any<bool>(k_ChangeResolution));
-  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float64>>(k_Origin));
-  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float64>>(k_Spacing));
+  args.insert(SetImageGeomOriginScalingFilter::k_Origin_Key, std::make_any<std::vector<float32>>(k_Origin));
+  args.insert(SetImageGeomOriginScalingFilter::k_Spacing_Key, std::make_any<std::vector<float32>>(k_Spacing));
 
   auto preflightResult = filter.preflight(dataStructure, args);
   SIMPLNX_RESULT_REQUIRE_VALID(preflightResult.outputActions);
